fix comments remarks
