+++ azurerm/cmd/doc.go
@@ -1,0 +1,4 @@
+// Package main allows to generate methods to get resources from Azure SDK
+// It uses the template defined in template.go and the resource function params defined in generate.go, to create a file in ./reader_generated.go
+// These methods can then be used to list the resources of Azure.
+package main
